+++ node/clients/search-resolver.ts
@@ -3,7 +3,13 @@ import { pathOr } from "ramda";
 
 export class SearchResolver extends AppGraphQLClient {
   constructor(context: IOContext, options?: InstanceOptions) {
-    super("vtex.search-resolver", context, options);
+    super("vtex.search-resolver", context, {
+      ...options,
+      headers: {
+        ...(options?.headers ?? {}),
+        'x-vtex-provider': 'vtex.search-graphql@0.43.0'
+      },
+    });
   }
 
   public productsById = async (ids: string[]) => {